requirements.txt
Flask==3.0.3
pymongo==4.2.0
dnspython==2.2.1
python-dotenv==1.0.0
